Major update (V0.4.0)
App interface has been updated with additional functionality.
Users can now subset data based on data, manual selected time range, and automatically removed data outside of the physical range module (values that dont make sense in the physical world - values sourced from Ameriflux technical documents; more info provided in the (?) at the top of the app)

If users utilize the PRM function (also available as its own function outside of the app!), the export feature will include the function from fluxtools in the r script with the generated code for manual removals - and will provide a second script for the specific PRM removals (these are often very large so I decided to keep it separate). A summary .csv is provided for each variable, number of rows for that variable set to NA, and the percentage overall removed (set to NA), along with another .csv for which values specifically were removed.